Moves eth2 staking to pinia
